+++ crates/macros/src/function.rs
@@ -1,11 +1,12 @@
-use std::{borrow::Cow, collections::HashMap};
+use std::collections::HashMap;
 
 use darling::{FromMeta, ToTokens};
 use proc_macro2::{Span, TokenStream};
-use quote::quote;
+use quote::{format_ident, quote};
 use syn::{spanned::Spanned, AttributeArgs, FnArg, Ident, ItemFn, Lit, PatType, Type};
 
-use crate::{bail, syn_ext::DropLifetimes, Result};
+use crate::prelude::*;
+use crate::syn_ext::DropLifetimes;
 
 #[derive(Debug, Default, FromMeta)]
 #[darling(default)]
@@ -37,12 +38,12 @@ pub fn parser(opts: AttributeArgs, input: ItemFn) -> Result<TokenStream> {
         Ok(opts) => opts,
         Err(e) => bail!("Failed to parse attribute options: {:?}", e),
     };
-    let args = Args::parse_from_fnargs(input.sig.inputs.iter(), &opts.defaults)?;
+    let args = Args::parse_from_fnargs(input.sig.inputs.iter(), opts.defaults)?;
     if let Some(ReceiverArg { span, .. }) = args.receiver {
         bail!(span => "Receiver arguments are invalid on PHP functions. See `#[php_impl]`.");
     }
-    let func = Function::parse(opts.name.as_deref(), &input, args)?;
-    let function_impl = func.php_function_impl(opts.optional.as_ref())?;
+    let func = Function::parse(&input.sig, opts.name, args, opts.optional)?;
+    let function_impl = func.php_function_impl()?;
 
     Ok(quote! {
         #input
@@ -50,36 +51,77 @@ pub fn parser(opts: AttributeArgs, input: ItemFn) -> Result<TokenStream> {
     })
 }
 
-struct Function<'a> {
-    name: Cow<'a, Ident>,
-    args: Args<'a>,
-    output: Option<&'a Type>,
+#[derive(Debug)]
+pub struct Function<'a> {
+    /// Identifier of the Rust function associated with the function.
+    pub ident: &'a Ident,
+    /// Name of the function in PHP.
+    pub name: String,
+    /// Function arguments.
+    pub args: Args<'a>,
+    /// Function outputs.
+    pub output: Option<&'a Type>,
+    /// The first optional argument of the function.
+    pub optional: Option<Ident>,
+}
+
+#[derive(Debug)]
+pub enum CallType<'a> {
+    Function,
+    Method {
+        class: &'a syn::Path,
+        receiver: MethodReceiver,
+    },
+}
+
+/// Type of receiver on the method.
+#[derive(Debug)]
+pub enum MethodReceiver {
+    /// Static method - has no receiver.
+    Static,
+    /// Class method, takes `&self` or `&mut self`.
+    Class,
+    /// Class method, takes `&mut ZendClassObject<Self>`.
+    ZendClassObject,
 }
 
 impl<'a> Function<'a> {
-    fn parse(name: Option<&str>, func: &'a ItemFn, args: Args<'a>) -> Result<Self> {
+    /// Parse a function.
+    ///
+    /// # Parameters
+    ///
+    /// * `sig` - Function signature.
+    /// * `name` - Function name in PHP land.
+    /// * `args` - Function arguments.
+    /// * `optional` - The ident of the first optional argument.
+    pub fn parse(
+        sig: &'a syn::Signature,
+        name: Option<String>,
+        args: Args<'a>,
+        optional: Option<Ident>,
+    ) -> Result<Self> {
         Ok(Self {
-            name: match name {
-                Some(name) => Cow::Owned(Ident::new(name, Span::call_site())),
-                None => Cow::Borrowed(&func.sig.ident),
-            },
+            ident: &sig.ident,
+            name: name.unwrap_or_else(|| sig.ident.to_string()),
             args,
-            output: match &func.sig.output {
+            output: match &sig.output {
                 syn::ReturnType::Default => None,
                 syn::ReturnType::Type(_, ty) => Some(&**ty),
             },
+            optional,
         })
     }
 
-    fn internal_ident(&self) -> Ident {
-        Ident::new(&format!("_internal_{}", &self.name), Span::call_site())
+    /// Generates an internal identifier for the function.
+    pub fn internal_ident(&self) -> Ident {
+        format_ident!("_internal_{}", &self.ident)
     }
 
-    fn php_function_impl(&self, optional: Option<&Ident>) -> Result<TokenStream> {
+    /// Generates the function builder for the function.
+    pub fn function_builder(&self, call_type: CallType) -> Result<TokenStream> {
+        let ident = self.ident;
         let name = &self.name;
-        let name_str = self.name.to_string();
-        let internal_ident = self.internal_ident();
-        let (required, not_required) = self.args.split_args(optional);
+        let (required, not_required) = self.args.split_args(self.optional.as_ref());
 
         // `handler` impl
         let required_arg_names: Vec<_> = required.iter().map(|arg| arg.name).collect();
@@ -90,7 +132,9 @@ impl<'a> Function<'a> {
             .iter()
             .map(TypedArg::arg_decleration)
             .collect::<Result<Vec<_>>>()?;
-        let arg_accessors = self.args.typed.iter().map(TypedArg::accessor);
+        let arg_accessors = self.args.typed.iter().map(|arg| {
+            arg.accessor(|e| quote! { #e.throw().expect("Failed to throw PHP exception.") })
+        });
 
         // `entry` impl
         let required_args = required
@@ -111,6 +155,96 @@ impl<'a> Function<'a> {
             }
         });
 
+        let result = match call_type {
+            CallType::Function => quote! {
+                let parse = ex.parser()
+                    #(.arg(&mut #required_arg_names))*
+                    .not_required()
+                    #(.arg(&mut #not_required_arg_names))*
+                    .parse();
+                if parse.is_err() {
+                    return;
+                }
+
+                #ident(#({#arg_accessors}),*)
+            },
+            CallType::Method { class, receiver } => {
+                let this = match receiver {
+                    MethodReceiver::Static => quote! {
+                        let parse = ex.parser();
+                    },
+                    MethodReceiver::ZendClassObject | MethodReceiver::Class => quote! {
+                        let (parse, this) = ex.parser_method::<#class>();
+                        let this = match this {
+                            Some(this) => this,
+                            None => {
+                                ::ext_php_rs::exception::PhpException::default("Failed to retrieve reference to `$this`".into())
+                                    .throw()
+                                    .unwrap();
+                                return;
+                            }
+                        };
+                    },
+                };
+                let call = match receiver {
+                    MethodReceiver::Static => {
+                        quote! { #class::#ident(#({#arg_accessors}),*) }
+                    }
+                    MethodReceiver::Class => quote! { this.#ident(#({#arg_accessors}),*) },
+                    MethodReceiver::ZendClassObject => {
+                        quote! { #class::#ident(this, #({#arg_accessors}),*) }
+                    }
+                };
+                quote! {
+                    #this
+                    let parse_result = parse
+                        #(.arg(&mut #required_arg_names))*
+                        .not_required()
+                        #(.arg(&mut #not_required_arg_names))*
+                        .parse();
+                    if parse_result.is_err() {
+                        return;
+                    }
+
+                    #call
+                }
+            }
+        };
+
+        Ok(quote! {
+            ::ext_php_rs::builders::FunctionBuilder::new(#name, {
+                ::ext_php_rs::zend_fastcall! {
+                    extern fn handler(
+                        ex: &mut ::ext_php_rs::zend::ExecuteData,
+                        retval: &mut ::ext_php_rs::types::Zval,
+                    ) {
+                        use ::ext_php_rs::convert::IntoZval;
+
+                        #(#arg_declerations)*
+                        let result = {
+                            #result
+                        };
+
+                        if let Err(e) = result.set_zval(retval, false) {
+                            let e: ::ext_php_rs::exception::PhpException = e.into();
+                            e.throw().expect("Failed to throw PHP exception.");
+                        }
+                    }
+                }
+                handler
+            })
+            #(.arg(#required_args))*
+            .not_required()
+            #(.arg(#not_required_args))*
+            #output
+        })
+    }
+
+    /// Generates a struct and impl for the `PhpFunction` trait.
+    pub fn php_function_impl(&self) -> Result<TokenStream> {
+        let internal_ident = self.internal_ident();
+        let builder = self.function_builder(CallType::Function)?;
+
         Ok(quote! {
             #[doc(hidden)]
             #[allow(non_camel_case_types)]
@@ -120,42 +254,71 @@ impl<'a> Function<'a> {
                 const FUNCTION_ENTRY: fn() -> ::ext_php_rs::error::Result<
                     ::ext_php_rs::zend::FunctionEntry
                 > = {
-                    ::ext_php_rs::zend_fastcall! {
-                        extern fn handler(
-                            ex: &mut ::ext_php_rs::zend::ExecuteData,
-                            retval: &mut ::ext_php_rs::types::Zval,
-                        ) {
-                            use ::ext_php_rs::convert::IntoZval;
-
-                            #(#arg_declerations)*
-                            let parse = ex.parser()
-                                #(.arg(&mut #required_arg_names))*
-                                .not_required()
-                                #(.arg(&mut #not_required_arg_names))*
-                                .parse();
-                            if parse.is_err() {
-                                return;
-                            }
-
-                            let result = #name(#({#arg_accessors}),*);
-                            if let Err(e) = result.set_zval(retval, false) {
-                                let e: ::ext_php_rs::exception::PhpException = e.into();
-                                e.throw().expect("Failed to throw PHP exception.");
-                            }
-                        }
-                    }
                     fn entry() -> ::ext_php_rs::error::Result<
                         ::ext_php_rs::zend::FunctionEntry
                     > {
-                        ::ext_php_rs::builders::FunctionBuilder::new(#name_str, handler)
-                            #(.arg(#required_args))*
-                            .not_required()
-                            #(.arg(#not_required_args))*
-                            #output
-                            .build()
+                        #builder.build()
                     }
                     entry
                 };
+            }
+        })
+    }
+
+    /// Returns a constructor metadata object for this function. This doesn't
+    /// check if the function is a constructor, however.
+    pub fn constructor_meta(&self, class: &syn::Path) -> Result<TokenStream> {
+        let ident = self.ident;
+        let (required, not_required) = self.args.split_args(self.optional.as_ref());
+        let required_args = required
+            .iter()
+            .map(TypedArg::arg_builder)
+            .collect::<Result<Vec<_>>>()?;
+        let not_required_args = not_required
+            .iter()
+            .map(TypedArg::arg_builder)
+            .collect::<Result<Vec<_>>>()?;
+
+        let required_arg_names: Vec<_> = required.iter().map(|arg| arg.name).collect();
+        let not_required_arg_names: Vec<_> = not_required.iter().map(|arg| arg.name).collect();
+        let arg_declerations = self
+            .args
+            .typed
+            .iter()
+            .map(TypedArg::arg_decleration)
+            .collect::<Result<Vec<_>>>()?;
+        let arg_accessors = self.args.typed.iter().map(|arg| {
+            arg.accessor(
+                |e| quote! { return ::ext_php_rs::class::ConstructorResult::Exception(#e); },
+            )
+        });
+
+        Ok(quote! {
+            ::ext_php_rs::class::ConstructorMeta {
+                constructor: {
+                    fn inner(ex: &mut ::ext_php_rs::zend::ExecuteData) -> ::ext_php_rs::class::ConstructorResult<#class> {
+                        #(#arg_declerations)*
+                        let parse = ex.parser()
+                            #(.arg(&mut #required_arg_names))*
+                            .not_required()
+                            #(.arg(&mut #not_required_arg_names))*
+                            .parse();
+                        if parse.is_err() {
+                            return ::ext_php_rs::class::ConstructorResult::ArgError;
+                        }
+                        #class::#ident(#({#arg_accessors}),*).into()
+                    }
+                    inner
+                },
+                build_fn: {
+                    fn inner(func: ::ext_php_rs::builders::FunctionBuilder) -> ::ext_php_rs::builders::FunctionBuilder {
+                        func
+                            #(.arg(#required_args))*
+                            .not_required()
+                            #(.arg(#not_required_args))*
+                    }
+                    inner
+                }
             }
         })
     }
@@ -172,7 +335,7 @@ pub struct TypedArg<'a> {
     pub name: &'a Ident,
     pub ty: &'a Type,
     pub nullable: bool,
-    pub default: Option<&'a Lit>,
+    pub default: Option<Lit>,
 }
 
 #[derive(Debug)]
@@ -184,7 +347,7 @@ pub struct Args<'a> {
 impl<'a> Args<'a> {
     pub fn parse_from_fnargs(
         args: impl Iterator<Item = &'a FnArg>,
-        defaults: &'a HashMap<Ident, Lit>,
+        mut defaults: HashMap<Ident, Lit>,
     ) -> Result<Self> {
         let mut result = Self {
             receiver: None,
@@ -213,7 +376,7 @@ impl<'a> Args<'a> {
                         name: ident,
                         ty: &**ty,
                         nullable,
-                        default: defaults.get(ident),
+                        default: defaults.remove(ident),
                     });
                 }
             }
@@ -294,9 +457,9 @@ impl<'a> TypedArg<'a> {
     }
 
     /// Get the accessor used to access the value of the argument.
-    fn accessor(&self) -> TokenStream {
+    fn accessor(&self, bail_fn: impl Fn(TokenStream) -> TokenStream) -> TokenStream {
         let name = self.name;
-        if let Some(default) = self.default {
+        if let Some(default) = &self.default {
             quote! {
                 #name.val().unwrap_or(#default)
             }
@@ -308,16 +471,16 @@ impl<'a> TypedArg<'a> {
                 #name.val()
             }
         } else {
+            let bail = bail_fn(quote! {
+                ::ext_php_rs::exception::PhpException::default(
+                    concat!("Invalid value given for argument `", stringify!(#name), "`.").into()
+                )
+            });
             quote! {
                 match #name.val() {
                     Some(val) => val,
                     None => {
-                        ::ext_php_rs::exception::PhpException::default(
-                            concat!("Invalid value given for argument `", stringify!(#name), "`.").into()
-                        )
-                        .throw()
-                        .expect("Failed to throw PHP exception.");
-                        return;
+                        #bail;
                     }
                 }
             }
