+++ crates/macros/src/lib.rs
@@ -3,7 +3,6 @@ mod constant;
 mod extern_;
 mod fastcall;
 mod function;
-mod helpers;
 mod impl_;
 mod method;
 mod module;
@@ -11,12 +10,6 @@ mod startup_function;
 mod syn_ext;
 mod zval;
 
-use std::{
-    collections::HashMap,
-    sync::{Mutex, MutexGuard},
-};
-
-use constant::Constant;
 use proc_macro::TokenStream;
 use proc_macro2::Span;
 use syn::{
@@ -26,31 +19,6 @@ use syn::{
 
 extern crate proc_macro;
 
-#[derive(Default, Debug)]
-struct State {
-    functions: Vec<function::Function>,
-    classes: HashMap<String, class::Class>,
-    constants: Vec<Constant>,
-    startup_function: Option<String>,
-    built_module: bool,
-}
-
-lazy_static::lazy_static! {
-    pub(crate) static ref STATE: StateMutex = StateMutex::new();
-}
-
-struct StateMutex(Mutex<State>);
-
-impl StateMutex {
-    pub fn new() -> Self {
-        Self(Mutex::new(Default::default()))
-    }
-
-    pub fn lock(&self) -> MutexGuard<State> {
-        self.0.lock().unwrap_or_else(|e| e.into_inner())
-    }
-}
-
 #[proc_macro_attribute]
 pub fn php_class(args: TokenStream, input: TokenStream) -> TokenStream {
     let args = parse_macro_input!(args as AttributeArgs);
@@ -58,7 +26,7 @@ pub fn php_class(args: TokenStream, input: TokenStream) -> TokenStream {
 
     match class::parser(args, input) {
         Ok(parsed) => parsed,
-        Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
+        Err(e) => e.to_compile_error(),
     }
     .into()
 }
@@ -69,8 +37,8 @@ pub fn php_function(args: TokenStream, input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as ItemFn);
 
     match function::parser(args, input) {
-        Ok((parsed, _)) => parsed,
-        Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
+        Ok(parsed) => parsed,
+        Err(e) => e.to_compile_error(),
     }
     .into()
 }
@@ -79,19 +47,14 @@ pub fn php_function(args: TokenStream, input: TokenStream) -> TokenStream {
 pub fn php_module(_: TokenStream, input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as ItemFn);
 
-    match module::parser(input) {
-        Ok(parsed) => parsed,
-        Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
-    }
-    .into()
+    module::parser(input).into()
 }
 
 #[proc_macro_attribute]
-pub fn php_startup(args: TokenStream, input: TokenStream) -> TokenStream {
-    let args = parse_macro_input!(args as AttributeArgs);
+pub fn php_startup(_: TokenStream, input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as ItemFn);
 
-    match startup_function::parser(Some(args), input) {
+    match startup_function::parser(input) {
         Ok(parsed) => parsed,
         Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
     }
@@ -105,7 +68,7 @@ pub fn php_impl(args: TokenStream, input: TokenStream) -> TokenStream {
 
     match impl_::parser(args, input) {
         Ok(parsed) => parsed,
-        Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
+        Err(e) => e.to_compile_error(),
     }
     .into()
 }
@@ -152,4 +115,55 @@ pub fn zend_fastcall(input: TokenStream) -> TokenStream {
         Err(e) => syn::Error::new(Span::call_site(), e).to_compile_error(),
     }
     .into()
+}
+
+#[proc_macro]
+pub fn wrap_function(input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(input as syn::Path);
+
+    match function::wrap(input) {
+        Ok(parsed) => parsed,
+        Err(e) => e.to_compile_error(),
+    }
+    .into()
+}
+
+macro_rules! err {
+    ($span:expr => $($msg:tt)*) => {
+        ::syn::Error::new(::syn::spanned::Spanned::span(&$span), format!($($msg)*))
+    };
+    ($($msg:tt)*) => {
+        ::syn::Error::new(::proc_macro2::Span::call_site(), format!($($msg)*))
+    };
+}
+
+/// Bails out of a function with a syn error.
+macro_rules! bail {
+    ($span:expr => $($msg:tt)*) => {
+        return Err($crate::err!($span => $($msg)*))
+    };
+    ($($msg:tt)*) => {
+        return Err($crate::err!($($msg)*))
+    };
+}
+
+pub(crate) use bail;
+pub(crate) use err;
+
+pub(crate) mod prelude {
+    pub(crate) trait OptionTokens {
+        fn option_tokens(&self) -> proc_macro2::TokenStream;
+    }
+
+    impl<T: quote::ToTokens> OptionTokens for Option<T> {
+        fn option_tokens(&self) -> proc_macro2::TokenStream {
+            match self {
+                Some(token) => quote::quote! { ::std::option::Option::Some(#token) },
+                None => quote::quote! { ::std::option::Option::None },
+            }
+        }
+    }
+
+    pub(crate) use crate::{bail, err};
+    pub(crate) type Result<T> = std::result::Result<T, syn::Error>;
 }
